Move existing scheduler to its own `strobe` module
