+++ shortfin/python/shortfin_apps/llm/components/scheduling/factory.py
@@ -3,14 +3,15 @@ import shortfin as sf
 from typing import List
 
 from .config import SchedulerConfig, SchedulingModes
+from .modes.strobe import StrobeScheduler
 from ..messages import LlmInferenceExecRequest
-from ..scheduler import Scheduler
+
 
 from shortfin_apps.utils import BatcherProcess
 
 
 class _ScheduleEngineImpl:
-    def __init__(self, scheduler: Scheduler) -> None:
+    def __init__(self, scheduler: StrobeScheduler) -> None:
         self.scheduler = scheduler
 
     def should_execute(
@@ -28,7 +29,7 @@ class _ScheduleEngineImpl:
 def _create_scheduler(config: SchedulerConfig) -> _ScheduleEngineImpl:
     if config.mode == SchedulingModes.STROBE:
         return _ScheduleEngineImpl(
-            scheduler=Scheduler(ideal_batch_size=config.ideal_batch_size),
+            scheduler=StrobeScheduler(ideal_batch_size=config.ideal_batch_size),
         )
 
     raise ValueError(f"Unsupported scheduling mode: {config.mode}")
